Move wait and commit functions to Log impl
